requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
jupyter==1.0.0
jupyterlab==4.0.5
lightgbm==4.0.0
xgboost==1.7.6
catboost==1.2
optuna==3.3.0
kaggle==1.5.16
tqdm==4.65.0
pyyaml==6.0.1
joblib==1.3.2
scipy==1.11.1
statsmodels==0.14.0
feature-engine==1.6.2
imbalanced-learn==0.11.0
shap==0.42.1
eli5==0.13.0
mlxtend==0.22.0
